requirements.txt
pandas==2.1.0
numpy==1.24.3
Faker==19.3.1
streamlit==1.28.0
plotly==5.18.0
nltk==3.8.1
wordcloud==1.9.2
scikit-learn==1.3.0
networkx==3.1
opencv-python==4.8.0.76
Pillow==10.0.0 
